# backend/openapi.yaml
openapi: 3.0.3
info:
  title: SE-Textile Backend API
  description: Dynamically generated OpenAPI spec built from registered Flask routes.
  version: 2.0.0
servers:
- url: http://127.0.0.1:5001
  description: Local development server
security:
- bearerAuth: []
paths:
...
  /api/v1/inventory/:
    get:
      tags:
      - inventory
      summary: Fetch all inventory products for a given shop.
      description: Fetch all inventory products for a given shop.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccess'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
      parameters:
      - name: shop_id
        in: query
        required: false
        schema:
          type: string
  /api/v1/inventory/delete:
    delete:
      tags:
      - inventory
      summary: Delete a product and its inventory entry.
      description: Delete a product and its inventory entry.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccess'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
      parameters:
      - name: product_id
        in: query
        required: false
        schema:
          type: string
  /api/v1/inventory/edit:
    post:
      tags:
      - inventory
      summary: Edit price or stock of a product.
      description: Edit price or stock of a product.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccess'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
      requestBody:
        required: false
        content:
          application/json: &id005
            schema:
              type: object
              properties:
                price:
                  type: string
                product_id:
                  type: string
                stock:
                  type: string
          application/x-www-form-urlencoded: *id005
  /api/v1/inventory/export:
    get:
      tags:
      - inventory
      summary: Download all inventory for a shop as Excel.
      description: Download all inventory for a shop as Excel.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccess'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
      parameters:
      - name: shop_id
        in: query
        required: false
        schema:
          type: string
  /api/v1/inventory/import:
    post:
      tags:
      - inventory
      summary: Upload inventory via CSV/Excel for a shop.
      description: Upload inventory via CSV/Excel for a shop.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccess'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericError'
      requestBody:
        required: false
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                shop_id:
                  type: string
                file:
                  type: string
                  format: binary
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    GenericSuccess:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
    GenericError:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        error:
          type: string
...